User api crud 복습2

+++ src/main/java/com/sparta/java02/domain/product/dto/ProductResponse.java
@@ -22,7 +22,7 @@ public class ProductResponse {
 
   String description;
 
-  BigDecimal price;
+  double price;
 
   Integer stock;
